Use direction/rotation to change wind icon
Added a function formatWindDirection
This takes wind speed and direction as either an abbreviation "NE" or rotation from north (90 etc). And generates rotation and direction.
If direction is "CALM" or speed is <1 km/h the calm icon is used (and rotation is set to zero) if direction is given and speed is >1 the wind_S icon is used and rotated to point in the given direction.