background layer changer
This adds a modal popup that lets you select some of the default mapbox 'background' layers. 
Note. Mapbox has no concept of background and layer objects. Instead it merges everything into one layer called the style. This means that in order to update the background style you have to re-connect all the layers.