Uodating Icons
Wanted to add some custom icons to the date selector buttons. The idea was that I could try and force the icons to be the same size to help fix the weird look. Turns out there's more to it than that. Also put some time into creating some missing icons. These were updated in cssipfont.